feat: refactor zustand store
