+++ utils/store/profile-store.ts
@@ -1,0 +1,13 @@
+import { create } from "zustand";
+
+type ProfileStore = {
+  image: string;
+  setImage: (newImage: string) => void;
+};
+
+export const useProfileStore = create<ProfileStore>((set) => ({
+  image: "",
+  setImage: (newImage) => {
+    set({ image: newImage });
+  },
+}));
