+++ utils/store/user-id-store.ts
@@ -1,17 +1,5 @@
 import { create } from "zustand";
 
-type ProfileStore = {
-  image: string;
-  setImage: (newImage: string) => void;
-};
-
-export const useProfileStore = create<ProfileStore>((set) => ({
-  image: "",
-  setImage: (newImage) => {
-    set({ image: newImage });
-  },
-}));
-
 type UserIdStore = {
   userId: string;
   setUserId: (newUserId: string) => void;
